Discovered tello bug #6 incorrect pathing
