+++ Tests/Controllers/Unit/DroneSimTests.cs
@@ -24,7 +24,7 @@ namespace Tests.Controllers.Unit
     public class DispatcherControllerTests
     {
         [Fact]
-        public async Task DroneSimShouldReturnTrue()
+        public async Task DroneSimDeliverShouldReturnTrue()
         {
             // Assumed to return an ok object result with ok as arg
 
@@ -46,7 +46,7 @@ namespace Tests.Controllers.Unit
         //     var response = await sim.InitializeRegistration(Constants.TestInitDroneRequest);
         //     response.Should().NotBeNull();
         //     response.Should().NotBeEquivalentTo(Constants.TestRecord.ToString());
-        // }
+        // } No longer exists
         // TODO: untestable, instantiates a gateway and overwrites any mocking
         // [Fact]
         // public async Task DroneSimShouldReturnOkStartDrone()
@@ -75,69 +75,5 @@ namespace Tests.Controllers.Unit
         //     response.Should().NotBeEquivalentTo(ExpectedHttp);
         // } NO longer exists
 
-        [Fact]
-        public void drone_should_have_destination_in_route()
-        {
-            //TODO: Bug #5 drone moves in Latitude and Longitude direction indefinitely
-            var drone = new Drone(Constants.TestRecord, new DroneToDispatchGateway());
-            var route = drone.GetRoute();
-            route.Should()
-                .NotBeNull();
-            route.Should()
-                .ContainEquivalentOf(Constants.Destination);
-        }
-
-        [Fact]
-        public void drone_should_have_start_in_route()
-        {
-
-            //TODO: Bug #5 drone moves in Latitude and Longitude direction indefinitely
-            var drone = new Drone(Constants.TestRecord, new DroneToDispatchGateway());
-            var route = drone.GetRoute();
-            route.Should()
-                .NotBeNull();
-            route.Should()
-                .ContainEquivalentOf(Constants.HomeLocation);
-        }
-
-        [Fact]
-        public void TestGetRouteAllPositiveNumbers()
-        {
-
-            //TODO: Bug #5 drone moves in Latitude and Longitude direction indefinitely
-            var drone = new Drone(Constants.TestRecord, new DroneToDispatchGateway());
-            var route = drone.GetRoute();
-            route.Should()
-                .NotBeNull();
-            foreach (var geoLocation in route)
-            {
-                geoLocation.Latitude.Should()
-                    .BeGreaterThanOrEqualTo(0m);
-                geoLocation.Longitude.Should()
-                    .BeGreaterThanOrEqualTo(0m);
-            }
-        }
-
-        [Fact]
-        public void TestGetRouteAllNegativeNumbers()
-        {
-
-            // Bug #5 drone moves in Latitude and Longitude direction indefinitely
-            var mockedDroneGatewaySetup = new Mock<IDroneToDispatcherGateway>();
-            mockedDroneGatewaySetup.Setup(x => x.PatchDroneStatus(It.IsAny<DroneStatusUpdateRequest>()))
-                .Returns(Task.FromResult<BaseDto>(Constants.TestDroneStatusUpdateRequest));
-            var mockedDroneGateway = mockedDroneGatewaySetup.Object;
-            var drone = new Drone(Constants.TestRecord, mockedDroneGateway as DroneToDispatchGateway);
-            var route = drone.GetRoute();
-            route.Should()
-                .NotBeNull();
-            foreach (var geoLocation in route)
-            {
-                geoLocation.Latitude.Should()
-                    .BeLessThanOrEqualTo(0.0m);
-                geoLocation.Longitude.Should()
-                    .BeLessThanOrEqualTo(0.0m);
-            }
-        }
     }
 }
